Create helper functions for the populate cards functions
